+++ R/slopeChart.R
@@ -161,7 +161,7 @@ wjp_slope <- function(
       geom_text(aes(y       = target_var,
                     x       = labpos,
                     label   = labels_var),
-                family      = "Lato Full",
+                family      = wjp_font_family(),
                 fontface    = "bold",
                 size        = 3.514598,
                 show.legend = FALSE)
@@ -178,7 +178,7 @@ wjp_slope <- function(
       ggrepel::geom_text_repel(mapping = aes(y     = target_var,
                                              x     = labpos,
                                              label = labels_var),
-                               family      = "Lato Full",
+                               family      = wjp_font_family(),
                                fontface    = "bold",
                                size        = 3.514598,
                                show.legend = FALSE,